RATEST-223: Migrate view, edit, and delete elements from Clinician Facing Patient Dashboard Page to Patient Visits Dashboard Page.
